Toughword_Analyser/requirements.txt
PyMuPDF==1.23.7
langchain==0.1.13
openai==1.3.5
tiktoken==0.6.0
python-dotenv==1.0.0
